Invention 11: proofreading
Measure 5, beat 1, right hand: The first sixteenth note of this beat
should be D instead of B-flat. Source1 and Source2 (manuscript) have the
D, but Source3 has the B-flat. Consulting several other IMSLP editions,
the D is much more common, so I've decided to follow the manuscript as
authority for this note.