Invention 7: proofreading
Measure 5, beats 2/3/4, left hand: The notes for these three beats
should be transposed up by a third (within the key).

Measure 12, beat 2, left hand: The two notes of this beat should be
swapped. That is, the D occurs before the E. Justification: The source2
manuscript (IMSLP#467846) and source3 edition (IMSLP#498228) align with
this change, but the source1 edition (IMSLP#19938) disagrees. Checking a
few other IMSLP sources, it is more common to see the D occur before the
E instead of vice versa. Because of the fact that Bach's manuscript
unambiguously displays the D before E, I am inclined to treat it as most
authoratative in this case.