Invention 9: proofreading
Measure 9, beat 3, left hand: The third sixteenth note of this beat
should be D-natural instead of D-flat. Source1 and Source2 (manuscript)
agree that it should be natural, but Source3 adds the flat as a
parenthesized cautionary. This change reflects the manuscript and
Source1.

Measure 12, beat 3, left hand: The third sixteenth note of this beat
should be A-natural instead of A-flat. Source1 and Source2 (manuscript)
agree that it should be natural, but Source3 adds the flat as a
parenthesized cautionary. This change reflects the manuscript and
Source1.

Measure 31, beat 2, right hand: The last sixteenth note of this beat
should be E-natural instead of E-flat.