Invention 7: proofreading (round 2)
Measure 16, beat 4, right hand:

Bach's manuscript does not add an accidental to the first sixteenth note
(D) of beat 4, and this fact has led to various interpretations across
many editions of this work. In modern notational practice, assuming Bach
had not mistakenly omitted a sharp accidental (i.e. making the note a
D-sharp), a cautionary natural would be added to the D, because the
previous measure has a D-sharp in the same octave.

I think it most likely that Bach indeed meant for the note to be
D-natural and simply omitted a cautionary for the note; the repeated
2-beat right-hand figuration in measures 16-17 is descending from the
tonic (I) to dominant (V), and using a natural minor sequence (E -> D ->
C -> B) rather than harmonic minor (E -> D-sharp -> C -> B) seems more
appropriate for this specific musical passage.

Checking several other editions, on IMSLP and elsewhere, there is a
slight preference to the D-natural compared to D-sharp. Most urtext
editions omit an accidental, identical to the manuscript, implying a
D-natural.