Invention 15: proofreading (round 2)
Measure 7, beat 3, right hand:

The second sixteenth note in beat 3 of this measure, E, lacks an
accidental in Bach's manuscript, which creates some ambiguity, because
the previous beat has an E-sharp in a lower octave.

Similar to the situation described in my last change to Invention 7,
measure 16, this ambiguity has led to different interpretations across
editions.

My interpretation is that Bach simply omitted a cautionary accidental
for the E, despite the fact that an E-sharp appears in the previous beat
in a different octave. So, following the manuscript literally, the note
should be an E-natural.

From an analytic perspective, beats 3 and 4 of measure 7 mark a
transition to the beginning of a new section of the piece that
introduces descending-fifth sequences starting with B minor on beat 1 of
measure 8, so any E-sharps in these two beats would weaken this
transition to B minor.